Attach owner to Comments and Blogs

+++ app/routes/blog_routes.js
@@ -1,11 +1,11 @@
 const express = require('express')
-// const passport = require('passport')
+const passport = require('passport')
 const Blog = require('../models/blog')
 const customErrors = require('../../lib/custom_errors')
 const handle404 = customErrors.handle404
-// const requireOwnership = customErrors.requireOwnership
+const requireOwnership = customErrors.requireOwnership
 const removeBlanks = require('../../lib/remove_blank_fields')
-// const requireToken = passport.authenticate('bearer', { session: false })
+const requireToken = passport.authenticate('bearer', { session: false })
 const router = express.Router()
 const Comment = require('../models/comment')
 
@@ -28,8 +28,8 @@ router.get('/blogs/:id', (req, res, next) => {
 })
 
 // CREATE
-router.post('/blogs', (req, res, next) => {
-  // req.body.example.owner = req.user.id
+router.post('/blogs', requireToken, (req, res, next) => {
+  req.body.blog.owner = req.user.id
   Blog.create(req.body.blog)
     .then(blog => {
       res.status(201).json({ blog: blog.toObject() })
@@ -38,11 +38,12 @@ router.post('/blogs', (req, res, next) => {
 })
 
 // UPDATE
-router.patch('/blogs/:id', removeBlanks, (req, res, next) => {
+router.patch('/blogs/:id', requireToken, removeBlanks, (req, res, next) => {
+  delete req.body.blog.owner
   Blog.findById(req.params.id)
     .then(handle404)
     .then(blog => {
-      // requireOwnership(req, example)
+      requireOwnership(req, blog)
       return blog.update(req.body.blog)
     })
     .then(() => res.sendStatus(204))
@@ -50,11 +51,11 @@ router.patch('/blogs/:id', removeBlanks, (req, res, next) => {
 })
 
 // DESTROY
-router.delete('/blogs/:id', (req, res, next) => {
+router.delete('/blogs/:id', requireToken, (req, res, next) => {
   Blog.findById(req.params.id)
     .then(handle404)
     .then(blog => {
-      // requireOwnership(req, example)
+      requireOwnership(req, blog)
       const id = blog
       Comment.deleteMany({blog: id}, (err, res) => { if (err) throw err })
       blog.remove()
